LayoutTests/http/tests/navigation/ping-attribute/secure-anchor-cross-origin-UpgradeMixedContent-expected.txt

CONSOLE MESSAGE: [blocked] The page at https://127.0.0.1:8443/navigation/ping-attribute/resources/secure-anchor-cross-origin.html requested insecure content from http://localhost:8000/navigation/resources/save-ping.py?test=secure-anchor-cross-origin. This content was blocked and must be served over HTTPS.

Ping not sent
